version5.2
1.数据库的小改动，增加了一些测试数据
2.增加了房源查询列表功能
3.完善了权限系统

+++ src/main/java/cn/dao/FollowupClientMapper.java
@@ -5,37 +5,37 @@ import cn.entity.FollowupClient;
 public interface FollowupClientMapper {
     /**
      *
-     * @mbggenerated 2017-05-18
+     * @mbggenerated 2017-05-20
      */
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     /**
      *
-     * @mbggenerated 2017-05-18
+     * @mbggenerated 2017-05-20
      */
     int insert(FollowupClient record);
 
     /**
      *
-     * @mbggenerated 2017-05-18
+     * @mbggenerated 2017-05-20
      */
     int insertSelective(FollowupClient record);
 
     /**
      *
-     * @mbggenerated 2017-05-18
+     * @mbggenerated 2017-05-20
      */
-    FollowupClient selectByPrimaryKey(Integer id);
+    FollowupClient selectByPrimaryKey(Long id);
 
     /**
      *
-     * @mbggenerated 2017-05-18
+     * @mbggenerated 2017-05-20
      */
     int updateByPrimaryKeySelective(FollowupClient record);
 
     /**
      *
-     * @mbggenerated 2017-05-18
+     * @mbggenerated 2017-05-20
      */
     int updateByPrimaryKey(FollowupClient record);
 }
